moved logic to scripts.js
